Removed some more compiler warnings in Xcode

+++ App/inc/IsGISAXSTools.h
@@ -27,7 +27,7 @@ class TH2D;
 class IsGISAXSTools {
 public:
     struct AxisStructure {
-        int nbins;
+        size_t nbins;
         std::vector<double> xbins;
         std::string name;
     };
