+++ App/src/TestIsGISAXS9.cpp
@@ -67,7 +67,7 @@ void TestIsGISAXS9::finalise()
     this_files.push_back(Utils::FileSystem::GetHomePath()+"./Examples/IsGISAXS_examples/ex-9/this_pyramid_Z0.ima");
     this_files.push_back(Utils::FileSystem::GetHomePath()+"./Examples/IsGISAXS_examples/ex-9/this_pyramid_Z45.ima");
 
-    int ncomparison = m_results.size();
+    int ncomparison = (int)m_results.size();
 
     TCanvas *canvaces[2];
     canvaces[0] = DrawHelper::instance().createAndRegisterCanvas("TestIsGISAXS9_c1", "Pyramid DWBA formfactor");
